# Conflicts:
#	ext/jt/jt-1078-server-dashboard-ui/package.json

+++ debug/xtream-codec-core-debug/src/main/java/io/github/hylexus/xtream/debug/codec/core/demo04/DemoJt808Msg0100V2019.java
@@ -51,7 +51,7 @@ public class DemoJt808Msg0100V2019 extends BaseJt808Msg {
 
     // 6. [75]   BYTE    车牌颜色
     @Preset.RustStyle.u8
-    private byte color;
+    private short color;
 
     /**
      * 7. [76,n)   String    车辆标识
