+++ ext/jt/jt-808-server-spring-boot-starter-reactive/src/main/java/io/github/hylexus/xtream/codec/ext/jt808/builtin/messages/request/BuiltinMessage0500.java
@@ -39,7 +39,7 @@ public class BuiltinMessage0500 {
      * 对应的车辆控制消息的流水号
      */
     @Preset.JtStyle.Word(desc = "应答流水号")
-    private short flowId;
+    private int flowId;
 
     /**
      * 位置信息汇报消息体
